Update index.js to generate the 'home' page
To generate the contents of the 'home' page via the imported function
from home.js, index.js has been updated to now append the result from
the imported function to the DOM. Now, when the website loads, the
'home' page content is automatically generated and displayed when
index.js runs.

Consequently, the template.html has been updated to remove the temporary
HTML that was there previously for the purposes of styling the look
of the 'home' page.