Add styling to the 'contact us' page
To style the 'contact us' page of the website, temporary HTML has been added
to template.html. This will be removed once JS has been added that will
dynamically generate the content of the 'contact us' page.

Additionally, the '--default-font' variable has been changed to
'standard-font', along with a 'ramen-font' variable. With the previous
configuration, the implementation of the 'contact us' page would have
led to the font of the input element's inheriting the custom font used
for the website design, which is not the desired effect, hence the
change.