Update index.js to generate 'contact us' page
To generate the contents of the 'contact us' page, index.js was updated
to import and run a function from contact-us.js that generates the
contents of said page, and append it to the DOM. For now, when the
website loads, the 'contact us' page is automatically generated when
index.js runs.

With this, the template.html has been updated to remove the temporary
HTML that was there previously for the purposes of styling the look of
the 'contact us' page.