Refactor createHeading fn, add missing semicolons
The 'createHeading' function was intended to be used to create and return
a heading element. But because it was only creating and returning a
specific heading element type, this led to other functions not being
able to use the function, even though it is repeating what the
'createHeading' function does but with a different heading element.
Thus, the 'createHeading' function was refactored to take in different
heading element types in order to reduce the repetition of code.