Add styling to the 'menu' page
To style the 'menu' page of the website, temporary HTML has been added
to template.html. This will be removed once JS has been added that will
dynamically generate the content of the 'menu' page.

To prevent the page from automatically generating the 'home' page of
the website and overwriting the temporary HTML, the call to run the
'initContent' function from index.js has been removed temporarily.