Update index.js to generate 'menu' page
To generate the contents of the 'menu' page, index.js was updated
to import and run a function from menu.js that generates the
contents of said page, and append it to the DOM. This will occur when
the user clicks on the 'menu' button on the nav bar.

With this, the template.html has been updated to remove the temporary
HTML that was there previously for the purposes of styling the look of
the 'menu' page.