Add 'attributions.js' file
To generate the content for the 'attributions' modal dynamically,
the 'attributions.js' script was created and added to the 'js' folder.
The aim is for the index.js file to import a function from
'attributions.js' that will generate a modal containing all of the
attributions for the assets used on the website.