Update index.js to generate 'attributions' modal
To generate the contents of the 'attributions' modal, index.js was
updated to import and run a function from attributions.js. This
function generates a modal containing attributions for all the assets
that were used on the website. The modal will display when clicking on
the 'attributions/credits' button located on the footer of the website.

With this, the template.html has been updated to remove the temporary
HTML that was there previously for the purposes of styling the look of
the 'attributions' modal.